Rename package currentPath to explorer
