+++ src/main/java/seedu/address/model/AddressBook.java
@@ -6,7 +6,7 @@ import java.io.File;
 import java.util.List;
 
 import javafx.collections.ObservableList;
-import seedu.address.model.currentpath.FileList;
+import seedu.address.model.explorer.FileList;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.tag.UniqueTagList;
