+++ src/main/java/seedu/address/model/explorer/CurrentPath.java
@@ -1,4 +1,4 @@
-package seedu.address.model.currentpath;
+package seedu.address.model.explorer;
 
 import static java.util.Objects.requireNonNull;
