--- src/test/java/seedu/address/model/currentPath/FileListStub.java
@@ -1,7 +1,0 @@
-package seedu.address.model.currentPath;
-
-import seedu.address.model.currentpath.FileList;
-
-public class FileListStub extends FileList {
-
-}
