+++ src/test/java/seedu/address/model/explorer/FileListStub.java
@@ -1,0 +1,5 @@
+package seedu.address.model.explorer;
+
+public class FileListStub extends FileList {
+
+}
